requirement.txt
attrs==22.1.0
black==22.1.0
blue==0.9.1
click==8.1.3
exceptiongroup==1.0.4
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pathspec==0.10.2
platformdirs==2.6.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.8.0
pyflakes==2.4.0
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
tomli==2.0.1
